+++ app/controllers/admin/payroll_gender_tasks_controller.rb
@@ -1,0 +1,29 @@
+class Admin::PayrollGenderTasksController < Admin::TasksController
+  def create
+    @claim_checking_tasks = ClaimCheckingTasks.new(@claim)
+    @task = @claim.tasks.build(check_params)
+    @claim.attributes = claim_params
+
+    if claim_and_task_saved?
+      redirect_to next_task_path
+    else
+      @tasks_presenter = @claim.policy::AdminTasksPresenter.new(@claim)
+      render current_task_name
+    end
+  end
+
+  private
+
+  def claim_params
+    params.require(:claim).permit(:payroll_gender)
+  end
+
+  def claim_and_task_saved?
+    ActiveRecord::Base.transaction do
+      @claim.save!(context: :"payroll-gender-task")
+      @task.save!
+    end
+  rescue ActiveRecord::RecordInvalid
+    false
+  end
+end